+++ Brainfuck translator/bf.c
@@ -4,20 +4,28 @@
 #include <stdlib.h>
 #include <string.h>
 
+enum {
+    FILE_ERR = 1,
+    MEMERR
+};
+
 int translation (const char* istr)
 {
     FILE *fp;
-    if ((fp = fopen("test.bf", "w")) == NULL)
-        goto ERROR;
+    if ((fp = fopen("output.bf", "w")) == NULL) {
+        fputs("Cannot open/write to file\n", stderr);
+        exit(FILE_ERR);
+    }
 
     char *ostr = NULL;
-    if ((ostr = (char*) malloc (10256)) == NULL)
-        goto ERROR;
+    if ((ostr = (char*) malloc (10256)) == NULL) {
+        fputs("Failed to allocate memory for output string\n", stderr);
+        exit(MEMERR);
+    }
 
-    ostr[0] = '<';
     unsigned int ipos = 0, opos = 0;
-    char prevchar = istr[0];
-    // int cell0, cell1, extra;
+    char prevchar = 0;
+    unsigned char cell0, cell1, extra;
 
     while(istr[ipos] != '\0') {
         // check for same char and input char
@@ -39,10 +47,6 @@ int translation (const char* istr)
         // check distance between previous char and incoming char
 
         if (prevchar-istr[ipos] < 13 && prevchar-istr[ipos] > 0) {
-            if (ostr[opos] == '<') {
-                opos++;
-                ostr[opos] = '>';
-            }
             for (char i = 0; i < prevchar-istr[ipos]; i++) {
                 opos++;
                 ostr[opos] = '-';
@@ -53,10 +57,6 @@ int translation (const char* istr)
             ipos++;
             continue;
         } else if (istr[ipos]-prevchar < 13 && istr[ipos]-prevchar > 0) {
-            if (ostr[opos] == '<') {
-                opos++;
-                ostr[opos] = '>';
-            }
             for (char i = 0; i < istr[ipos]-prevchar; i++) {
                 opos++;
                 ostr[opos] = '+';
@@ -67,46 +67,51 @@ int translation (const char* istr)
             ipos++;
             continue;
         }
-        /* 
+        // fall through if none works above
         // calculation
-        int check;
-        for (int i = 10; i > 1; i--) { // first run
-            if (prevchar > (int) istr[ipos] && (prevchar-(int) istr[ipos])%i == 0) {
-                cell0 = (prevchar-(int) istr[ipos])/i;
+        int check = 0;
+        for (char i = 10; i > 1; i--) { // first run
+            if (prevchar > istr[ipos] && (prevchar-istr[ipos])%i == 0) {
+                cell0 = (prevchar-istr[ipos])/i;
                 cell1 = i;
                 extra = 0;
                 break;
-            } else if (prevchar < (int) istr[ipos] && (prevchar-(int) istr[ipos])%i == 0) {
-                cell0 = ((int) istr[ipos]-prevchar)/i;
+            } else if (prevchar < istr[ipos] && (istr[ipos]-prevchar)%i == 0) {
+                cell0 = (istr[ipos]-prevchar)/i;
                 cell1 = i;
                 extra = 0;
                 break;
-            }
-            else if (i == 2) {
+            } else if (i == 2) 
                 check = 1;
-            }
         }
         if (check == 1) { // second run if true
-            for (int j = 10; j > 1; j--) {
-                if (prevchar > (int) istr[ipos] && (prevchar-(int) istr[ipos])%j == 1) {
-                    cell0 = (prevchar-(int) istr[ipos])/j;
+            for (char j = 10; j > 1; j--) {
+                if (prevchar > istr[ipos] && (prevchar-istr[ipos])%j == 1) {
+                    cell0 = (prevchar-istr[ipos])/j;
                     cell1 = j;
                     extra = 1;
                     break;
-                } else if (prevchar < (int) istr[ipos] && (prevchar-(int) istr[ipos])%j == 0) {
-                    cell0 = ((int) istr[ipos]-prevchar)/j;
+                } else if (prevchar < istr[ipos] && (istr[ipos]-prevchar)%j == 1) {
+                    cell0 = (istr[ipos]-prevchar)/j;
                     cell1 = j;
-                    extra = 0;
+                    extra = 1;
                     break;
                 }
             }
         }
         // writing to output string
-        // TODO: fix 
         
-        opos++;
-        ostr[opos] = '<';
+        if (ostr[opos] != '<' && opos != 0) { // return to cell 0
+            opos++;
+            ostr[opos] = '<';
+        }
         for (int i = 0; i < cell0; i++) {
+            if (opos == 0) {
+                ostr[opos] = '+';
+                i++;
+                goto SKIP;
+            }
+            SKIP:
             opos++;
             ostr[opos] = '+';
         }
@@ -115,15 +120,38 @@ int translation (const char* istr)
         opos++;
         ostr[opos] = '>';
         
-        if (prevchar > (int) istr[ipos]) {
-            for (int i = 0  ; i < cell1; i++) {
+        if (prevchar > istr[ipos]) { 
+            for (int i = 0; i < cell1; i++) {
                 opos++;
                 ostr[opos] = '-';
             }
-        } */
-        
-
-    ipos++;
+        } else if (prevchar < istr[ipos]) {
+            for (int i = 0; i < cell1; i++) {
+                opos++;
+                ostr[opos] = '+';
+            }
+        }
+        opos++;
+        ostr[opos] = '<';
+        opos++;
+        ostr[opos] = '-';
+        opos++;
+        ostr[opos] = ']';
+        opos++;
+        ostr[opos] = '>';
+        if (extra == 1) { // for prime numbers
+            if (prevchar > istr[ipos]) {
+                opos++;
+                ostr[opos] = '-';
+            } else if (prevchar < istr[ipos]) {
+                opos++;
+                ostr[opos] = '+';
+            }
+        }
+        opos++;
+        ostr[opos] = '.';
+        prevchar = istr[ipos];
+        ipos++;
     }
     opos++;
     ostr[opos] = '\0';
@@ -132,12 +160,6 @@ int translation (const char* istr)
     fclose(fp);
     free(ostr);
     return 0;
-
-    // will never fall through unless called by error checking routine
-
-    ERROR:
-    fprintf(stderr, "Memory error! Not enough memory...\n");
-    exit(1); 
 }
