+++ Brainfuck translator/old code/impbrainfuck.c
