--- Brainfuck translator/output.bf
@@ -1,1 +1,0 @@
-<>..+++.
